% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_trajectories.R
\name{.check_antimeridian}
\alias{.check_antimeridian}
\title{Check for antimeridian crossing and warn}
\usage{
.check_antimeridian(df)
}
\arguments{
\item{df}{data.frame with traj_id and x columns.}
}
\value{
NULL (invisible). May issue a warning.
}
\description{
Check for antimeridian crossing and warn
}
\keyword{internal}
