% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.check_class}
\alias{.check_class}
\title{Validate S3 class of workflow objects}
\usage{
.check_class(x, expected_class, next_function)
}
\arguments{
\item{x}{The object to check.}

\item{expected_class}{The expected S3 class name (character).}

\item{next_function}{The function that should have produced the object
(character), used in the error message.}
}
\value{
\code{NULL} (invisible). Throws \code{stop()} on class mismatch.
}
\description{
Checks that the input object has the expected S3 class and throws an
informative error pointing the user to the correct function.
}
\keyword{internal}
