% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_along_track_signed}
\alias{.cpp_along_track_signed}
\title{Compute signed along-track distance (C++)}
\usage{
.cpp_along_track_signed(plon, plat, lon1, lat1, lon2, lat2)
}
\arguments{
\item{plon, plat}{Point coordinates in degrees}

\item{lon1, lat1, lon2, lat2}{Great circle endpoints in degrees}
}
\value{
Signed along-track distance in meters (double)
}
\description{
Compute signed along-track distance (C++)
}
\keyword{internal}
