% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_cross_track_dist}
\alias{.cpp_cross_track_dist}
\title{Compute cross-track distance from point to great circle (C++)}
\usage{
.cpp_cross_track_dist(plon, plat, lon1, lat1, lon2, lat2)
}
\arguments{
\item{plon, plat}{Point coordinates in degrees}

\item{lon1, lat1, lon2, lat2}{Great circle endpoints in degrees}
}
\value{
Absolute cross-track distance in meters (double)
}
\description{
Compute cross-track distance from point to great circle (C++)
}
\keyword{internal}
