% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_d_angle_euc}
\alias{.cpp_d_angle_euc}
\title{Compute euclidean angle distance between two segments (C++)}
\usage{
.cpp_d_angle_euc(si_x, si_y, ei_x, ei_y, sj_x, sj_y, ej_x, ej_y)
}
\arguments{
\item{si_x, si_y, ei_x, ei_y}{Coordinates of segment i}

\item{sj_x, sj_y, ej_x, ej_y}{Coordinates of segment j}
}
\value{
Angle distance (double)
}
\description{
Compute euclidean angle distance between two segments (C++)
}
\keyword{internal}
