% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_d_par_sph}
\alias{.cpp_d_par_sph}
\title{Compute spherical parallel distance between two segments (C++)}
\usage{
.cpp_d_par_sph(lon_si, lat_si, lon_ei, lat_ei, lon_sj, lat_sj, lon_ej, lat_ej)
}
\arguments{
\item{lon_si, lat_si, lon_ei, lat_ei}{Segment i endpoints in degrees}

\item{lon_sj, lat_sj, lon_ej, lat_ej}{Segment j endpoints in degrees}
}
\value{
Parallel distance in meters (double)
}
\description{
Compute spherical parallel distance between two segments (C++)
}
\keyword{internal}
