% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_haversine_dist}
\alias{.cpp_haversine_dist}
\title{Compute haversine distance between two points (C++)}
\usage{
.cpp_haversine_dist(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1, lat1}{Coordinates of point 1 in degrees}

\item{lon2, lat2}{Coordinates of point 2 in degrees}
}
\value{
Distance in meters (double)
}
\description{
Compute haversine distance between two points (C++)
}
\keyword{internal}
