% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_initial_bearing}
\alias{.cpp_initial_bearing}
\title{Compute initial bearing between two points (C++)}
\usage{
.cpp_initial_bearing(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1, lat1}{Coordinates of point 1 in degrees}

\item{lon2, lat2}{Coordinates of point 2 in degrees}
}
\value{
Bearing in degrees [0, 360) (double)
}
\description{
Compute initial bearing between two points (C++)
}
\keyword{internal}
