% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_mdl_costnopar}
\alias{.cpp_mdl_costnopar}
\title{Compute MDL cost of no-partitioning (costnopar) for testing}
\usage{
.cpp_mdl_costnopar(sub_x, sub_y, method)
}
\arguments{
\item{sub_x}{Numeric vector of point x-coordinates.}

\item{sub_y}{Numeric vector of point y-coordinates.}

\item{method}{Integer: 0 for euclidean, 1 for haversine.}
}
\value{
The MDL costnopar value (double).
}
\description{
Compute MDL cost of no-partitioning (costnopar) for testing
}
\keyword{internal}
