% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_traclus_dist_euc}
\alias{.cpp_traclus_dist_euc}
\title{Compute weighted total euclidean distance between two segments (C++)}
\usage{
.cpp_traclus_dist_euc(
  si_x,
  si_y,
  ei_x,
  ei_y,
  sj_x,
  sj_y,
  ej_x,
  ej_y,
  w_perp,
  w_par,
  w_angle,
  eps
)
}
\arguments{
\item{si_x, si_y, ei_x, ei_y}{Coordinates of segment i}

\item{sj_x, sj_y, ej_x, ej_y}{Coordinates of segment j}

\item{w_perp, w_par, w_angle}{Distance component weights}

\item{eps}{Epsilon threshold for early termination}
}
\value{
Weighted total distance (double)
}
\description{
Compute weighted total euclidean distance between two segments (C++)
}
\keyword{internal}
