% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_traclus_dist_sph}
\alias{.cpp_traclus_dist_sph}
\title{Compute weighted total spherical distance between two segments (C++)}
\usage{
.cpp_traclus_dist_sph(
  lon_si,
  lat_si,
  lon_ei,
  lat_ei,
  lon_sj,
  lat_sj,
  lon_ej,
  lat_ej,
  w_perp,
  w_par,
  w_angle,
  eps
)
}
\arguments{
\item{lon_si, lat_si, lon_ei, lat_ei}{Segment i endpoints in degrees}

\item{lon_sj, lat_sj, lon_ej, lat_ej}{Segment j endpoints in degrees}

\item{w_perp, w_par, w_angle}{Distance component weights}

\item{eps}{Epsilon threshold for early termination}
}
\value{
Weighted total distance in meters (double)
}
\description{
Compute weighted total spherical distance between two segments (C++)
}
\keyword{internal}
