% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_trajectories.R
\name{.group_by_traj_id}
\alias{.group_by_traj_id}
\title{Group data by traj_id preserving original point order within groups}
\usage{
.group_by_traj_id(df)
}
\arguments{
\item{df}{data.frame with traj_id column.}
}
\value{
data.frame reordered by traj_id groups.
}
\description{
Group data by traj_id preserving original point order within groups
}
\keyword{internal}
