% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_along_track_signed}
\alias{.r_along_track_signed}
\title{R reference: signed along-track distance (meters)}
\usage{
.r_along_track_signed(p, a, b)
}
\arguments{
\item{p}{Point as c(lon, lat) in degrees.}

\item{a}{Start of great circle as c(lon, lat).}

\item{b}{End of great circle as c(lon, lat).}
}
\value{
Signed along-track distance in meters (double).
}
\description{
R reference: signed along-track distance (meters)
}
\keyword{internal}
