% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_bearing}
\alias{.r_bearing}
\title{R reference: initial bearing (forward azimuth) in degrees [0, 360)}
\usage{
.r_bearing(p1, p2)
}
\arguments{
\item{p1}{Point 1 as c(lon, lat) in degrees.}

\item{p2}{Point 2 as c(lon, lat) in degrees.}
}
\value{
Bearing in degrees (double).
}
\description{
R reference: initial bearing (forward azimuth) in degrees [0, 360)
}
\keyword{internal}
