% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_cross_track}
\alias{.r_cross_track}
\title{R reference: cross-track distance (absolute, meters)}
\usage{
.r_cross_track(p, a, b)
}
\arguments{
\item{p}{Point as c(lon, lat) in degrees.}

\item{a}{Start of great circle as c(lon, lat).}

\item{b}{End of great circle as c(lon, lat).}
}
\value{
Absolute cross-track distance in meters (double).
}
\description{
R reference: cross-track distance (absolute, meters)
}
\keyword{internal}
