% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_d_par_euc}
\alias{.r_d_par_euc}
\title{R reference: euclidean parallel distance (Paper Def. 2)}
\usage{
.r_d_par_euc(si, ei, sj, ej)
}
\arguments{
\item{si, ei, sj, ej}{Segment endpoints (each length-2 numeric).}
}
\value{
Parallel distance (double).
}
\description{
R reference: euclidean parallel distance (Paper Def. 2)
}
\keyword{internal}
