% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_d_par_sph}
\alias{.r_d_par_sph}
\title{R reference: spherical parallel distance}
\usage{
.r_d_par_sph(si, ei, sj, ej)
}
\arguments{
\item{si, ei, sj, ej}{Segment endpoints (each c(lon, lat)).}
}
\value{
Parallel distance in meters (double).
}
\description{
R reference: spherical parallel distance
}
\keyword{internal}
