% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_d_perp_sph}
\alias{.r_d_perp_sph}
\title{R reference: spherical perpendicular distance}
\usage{
.r_d_perp_sph(si, ei, sj, ej)
}
\arguments{
\item{si, ei, sj, ej}{Segment endpoints (each c(lon, lat)).}
}
\value{
Perpendicular distance in meters (double).
}
\description{
R reference: spherical perpendicular distance
}
\keyword{internal}
