% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_haversine}
\alias{.r_haversine}
\title{R reference: haversine distance in meters}
\usage{
.r_haversine(p1, p2)
}
\arguments{
\item{p1}{Point 1 as c(lon, lat) in degrees.}

\item{p2}{Point 2 as c(lon, lat) in degrees.}
}
\value{
Distance in meters (double).
}
\description{
R reference: haversine distance in meters
}
\keyword{internal}
