% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_point_to_line_dist}
\alias{.r_point_to_line_dist}
\title{R reference: perpendicular distance from point to LINE}
\usage{
.r_point_to_line_dist(p, ls, le)
}
\arguments{
\item{p}{Point (length-2 numeric).}

\item{ls}{Line start (length-2 numeric).}

\item{le}{Line end (length-2 numeric).}
}
\value{
Perpendicular distance (double).
}
\description{
R reference: perpendicular distance from point to LINE
}
\keyword{internal}
