% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_distances.R
\name{.r_seg_len_euc}
\alias{.r_seg_len_euc}
\title{R reference: euclidean segment length}
\usage{
.r_seg_len_euc(s, e)
}
\arguments{
\item{s}{Start point (length-2 numeric).}

\item{e}{End point (length-2 numeric).}
}
\value{
Euclidean distance (double).
}
\description{
R reference: euclidean segment length
}
\keyword{internal}
