% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_trajectories.R
\name{.truncate_ids}
\alias{.truncate_ids}
\title{Truncate a vector of IDs for display in warnings}
\usage{
.truncate_ids(ids, max_show = 5)
}
\arguments{
\item{ids}{Character vector of IDs.}

\item{max_show}{Maximum number to show before truncating.}
}
\value{
Formatted string.
}
\description{
Truncate a vector of IDs for display in warnings
}
\keyword{internal}
