% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tc_estimate}
\alias{print.tc_estimate}
\title{Print method for tc_estimate objects}
\usage{
\method{print}{tc_estimate}(x, ...)
}
\arguments{
\item{x}{A \code{tc_estimate} object from \code{\link[=tc_estimate_params]{tc_estimate_params()}}.}

\item{...}{Further arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for tc_estimate objects
}
\examples{
\donttest{
trj <- tc_trajectories(traclus_toy,
  traj_id = "traj_id",
  x = "x", y = "y", coord_type = "euclidean"
)
parts <- tc_partition(trj)
est <- tc_estimate_params(parts)
print(est)
}
}
