% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.TRACLUS}
\alias{print.TRACLUS}
\alias{print.tc_trajectories}
\alias{print.tc_partitions}
\alias{print.tc_clusters}
\alias{print.tc_representatives}
\alias{print.tc_traclus}
\title{Print methods for TRACLUS objects}
\usage{
\method{print}{tc_trajectories}(x, ...)

\method{print}{tc_partitions}(x, ...)

\method{print}{tc_clusters}(x, ...)

\method{print}{tc_representatives}(x, ...)

\method{print}{tc_traclus}(x, ...)
}
\arguments{
\item{x}{A TRACLUS object.}

\item{...}{Further arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Compact summaries of TRACLUS workflow objects, printed via \code{\link[=cat]{cat()}}.
All methods return the object invisibly for pipe compatibility.
}
\examples{
trj <- tc_trajectories(traclus_toy,
  traj_id = "traj_id",
  x = "x", y = "y", coord_type = "euclidean"
)
print(trj)

}
\seealso{
Other workflow functions: 
\code{\link{plot.TRACLUS}},
\code{\link{summary.TRACLUS}},
\code{\link{tc_cluster}()},
\code{\link{tc_leaflet}()},
\code{\link{tc_partition}()},
\code{\link{tc_plot}()},
\code{\link{tc_represent}()},
\code{\link{tc_traclus}()},
\code{\link{tc_trajectories}()}
}
\concept{workflow functions}
