% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.TRACLUS}
\alias{summary.TRACLUS}
\alias{summary.tc_trajectories}
\alias{summary.tc_partitions}
\alias{summary.tc_clusters}
\alias{summary.tc_representatives}
\alias{summary.tc_traclus}
\alias{summary.tc_estimate}
\title{Summary methods for TRACLUS objects}
\usage{
\method{summary}{tc_trajectories}(object, ...)

\method{summary}{tc_partitions}(object, ...)

\method{summary}{tc_clusters}(object, ...)

\method{summary}{tc_representatives}(object, ...)

\method{summary}{tc_traclus}(object, ...)

\method{summary}{tc_estimate}(object, ...)
}
\arguments{
\item{object}{A TRACLUS object.}

\item{...}{Further arguments (ignored).}
}
\value{
\code{object}, invisibly.
}
\description{
Detailed summaries of TRACLUS workflow objects, printed via \code{\link[=cat]{cat()}}.
All methods return the object invisibly.
}
\examples{
trj <- tc_trajectories(traclus_toy,
  traj_id = "traj_id",
  x = "x", y = "y", coord_type = "euclidean"
)
summary(trj)

}
\seealso{
Other workflow functions: 
\code{\link{plot.TRACLUS}},
\code{\link{print.TRACLUS}},
\code{\link{tc_cluster}()},
\code{\link{tc_leaflet}()},
\code{\link{tc_partition}()},
\code{\link{tc_plot}()},
\code{\link{tc_represent}()},
\code{\link{tc_traclus}()},
\code{\link{tc_trajectories}()}
}
\concept{workflow functions}
