% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACLUS-package.R
\docType{package}
\name{TRACLUS-package}
\alias{TRACLUS}
\alias{TRACLUS-package}
\title{TRACLUS: Trajectory Clustering Using Line Segments}
\description{
Implements the TRACLUS (TRAjectory CLUStering) algorithm for partitioning,
clustering, and summarizing trajectory data. The algorithm works in three
phases: (1) partition trajectories into line segments using minimum
description length (MDL), (2) cluster segments using density-based spatial
clustering (DBSCAN), and (3) generate representative trajectories using a
sweep-line algorithm. Supports both euclidean and spherical (haversine)
distance calculations.
}
\references{
Lee, J.-G., Han, J., & Whang, K.-Y. (2007). Trajectory clustering:
A partition-and-group framework. In \emph{Proceedings of the 2007 ACM
SIGMOD International Conference on Management of Data} (pp. 593--604).
\doi{10.1145/1247480.1247546}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MartinHoblisch/TRACLUS}
  \item \url{https://martinhoblisch.github.io/TRACLUS}
  \item Report bugs at \url{https://github.com/MartinHoblisch/TRACLUS/issues}
}

}
\author{
\strong{Maintainer}: Martin Hoblisch \email{m.hoblisch@gmail.com}

}
\keyword{internal}
